+++ backend/tracking/utils.py
@@ -1,16 +1,9 @@
 import requests
 import xml.etree.ElementTree as ElementTree
 from datetime import datetime
-from django.conf import settings
+from core.constants import VALID_STATUS_CODES, REQUEST_HEADERS
 
-from catalog.models import BoardGame
-
-VALID_STATUS_CODES = [200, 202]
-BGG_TOKEN = settings.BGG_API_TOKEN
-REQUEST_HEADERS: dict[str, str] = {
-    "User-Agent": "QuestLog/0.1",
-    "Authorization": f"Bearer {BGG_TOKEN}"
-}
+from catalog.utils import get_bgg_board_game, get_existing_board_game
 
 
 def test():
@@ -18,37 +11,14 @@ def test():
     get_bgg_board_game(id)
 
 
-def get_bgg_board_game(bgg_id: int) -> BoardGame:
-    fetch_url: str = f"https://boardgamegeek.com/xmlapi2/thing?id={bgg_id}"
-    response: requests.Response = requests.get(url=fetch_url, headers=REQUEST_HEADERS)
-    print(response.status_code)
-    response_root: ElementTree.Element = ElementTree.fromstring(response.content)
-
-    # TODO: parse xml response to find the primary name and year published
-    for item in response_root.findall("item"):
-        primary_name: ElementTree.Element[str] = item.find('name[@type="primary"]')
-        print(primary_name.attrib.get("value"))
-    # TODO: create a new BoardGame record or use fallback name if API failes
-    # TODO: return the BoardGame instance
-
-
-def get_existing_board_game(bgg_id: int, fallback_name: str) -> BoardGame:
-    existing_board_game: BoardGame = BoardGame.objects.get(bgg_id=bgg_id)
-
-    if existing_board_game:
-        return existing_board_game
-    else:
-        get_bgg_board_game(bgg_id=bgg_id)
-
-
 def fetch_bgg_plays(user: str, bgg_username: str) -> tuple[bool, str]:
     fetch_url: str = f"https://boardgamegeek.com/xmlapi2/plays?username={bgg_username}"
-
     response: requests.Response = requests.get(url=fetch_url, headers=REQUEST_HEADERS)
 
     if response.status_code not in VALID_STATUS_CODES:
         return False, f"Failed to reach BGG API. Status: {response.status_code}"
 
+    # XML Root should be 'plays'
     response_root: ElementTree.Element = ElementTree.fromstring(response.content)
 
     sessions_created: int = 0
